Split discoverable logic into opt-in and opt-out cases.
Before this change logic was:
- if actor is discoverable then index content.
- if actor is not discoverable then check content for opt-in,
  e.g. opposite setting for indexable flag. If there is
  opt-in for _this_ content item then index it. Otherwise
  ignore content item.

The missing bit in this logic was scenario when actor
is opt-in in general but wants to avoid indexing of
particular content item.

This change addresses that case by mirroring logic for
opt-in:
- if _actor_ is discoverable then check _content_ item
  properties for indication of opt-out. If there is no any
  then indexing is fine.
- second decision logic branch stays the same:
  if _actor_ is not discoverable then check content for opt-in,
  If there is opt-in for _this content item_ then index it.
  Otherwise ignore it.

Also this change updates how `discoverable` flag is checked
for actor in case `@context` property does not declare it.
Now property is checked outside condition to verify existence
of property in schema. This seems to be applicable to some
LitePub instances only.